First Steps to implement aes client basd encryption and recognized there is work todo in indexProc

+++ setupDatabase.php
@@ -1,16 +1,21 @@
 <?php
 include 'sqlConnect.php';
 
+
+if(!$mysqli->query("DROP TABLE IF EXISTS user"))
+	die("Error 3: " . $mysqli->error);
+
+if(!$mysqli->query("Create Table user (userName varchar (64) NOT NULL, passWord  varchar (64) NOT NULL, email varchar (64) NOT NULL, regDate TIMESTAMP DEFAULT CURRENT_TIMESTAMP, PRIMARY KEY(userName)"); 
+
 if(!$mysqli->query("DROP TABLE IF EXISTS savedLoginData"))
-	die("Error" . $mysqli->error);
+	die("Error 1: " . $mysqli->error);
 
-if(!$mysqli->query("Create Table savedLoginData (userName varchar(64) NOT NULL, foreignServiceName varchar(64), foreignServiceUserName varchar(64) NOT NULL, foreignServicePassWord varchar(64) NOT NULL, creationTimeStamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP, FOREIGN KEY (userName) REFERENCES user(userName), PRIMARY KEY (userName));"))
-	die("Error" . $mysqli->error);
-?>
+if(!$mysqli->query("Create Table savedLoginData (userName varchar(64) NOT NULL, foreignServiceName varchar(64) NOT NULL, foreignServiceUserName varchar(64) NOT NULL, foreignServicePassWord varchar(64) NOT NULL, creationTimeStamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP, FOREIGN KEY (userName) REFERENCES user(userName), PRIMARY KEY (userName));"))
+	die("Error 2: " . $mysqli->error);
 
-//same for new struct of user table
 
 //Drop old table if it exists
 
 //create new one with colmns ffrom README.md
+?>
